✂️Rename: Event 모듈 이동
성질상 common이 아닌 별도 디렉토리로 변경
